+++ resolwe/permissions/tests/__init__.py
@@ -1,0 +1,8 @@
+"""
+=======
+Testing
+=======
+
+.. automodule:: resolwe.permissions.tests.base
+
+"""